This simply breaks main.go into a handful of logical pieces.
There are zero functional changes in this commit.
